+++ eslint.config.js
@@ -1,0 +1,12 @@
+import { config } from '@joshunrau/eslint-config';
+
+export default config({
+  env: {
+    browser: false,
+    es2021: true,
+    node: true
+  },
+  typescript: {
+    enabled: true
+  }
+});