+++ src/__legacy/SongTagsSearch.ts
@@ -3,31 +3,31 @@ import type { MoreVideoDetails } from 'ytdl-core';
 
 import { YtdlMp3Error, removeParenthesizedText, userInput } from './utils';
 
-export interface SearchResult {
+export type SearchResult = {
   artistName: string;
-  trackName: string;
   artworkUrl100: string;
+  trackName: string;
 }
 
-export interface SearchData {
+export type SearchData = {
   resultCount: number;
   results: SearchResult[];
 }
 
-export interface AlbumArt {
+export type AlbumArt = {
+  description: string;
+  imageBuffer: Buffer;
   mime: string;
   type: {
     id: number;
     name: string;
   };
-  description: string;
-  imageBuffer: Buffer;
 }
 
-export interface SongTags {
-  title: string;
+export type SongTags = {
   artist: string;
   image: AlbumArt;
+  title: string;
 }
 
 export class SongTagsSearch {
@@ -44,22 +44,31 @@ export class SongTagsSearch {
   async search(verify = false): Promise<SongTags> {
     console.log(`Attempting to query iTunes API with the following search term: ${this.searchTerm}`);
     const searchResults = await this.fetchResults();
-    const result = verify ? await this.getVerifiedResult(searchResults) : searchResults[0];
+    const result = verify ? await this.getVerifiedResult(searchResults) : searchResults[0]!;
     const artworkUrl = result.artworkUrl100.replace('100x100bb.jpg', '600x600bb.jpg');
     const albumArt = await this.fetchAlbumArt(artworkUrl);
     return {
-      title: result.trackName,
       artist: result.artistName,
       image: {
+        description: 'Album Art',
+        imageBuffer: albumArt,
         mime: 'image/png',
         type: {
           id: 3,
-          name: 'front cover',
-        },
-        description: 'Album Art',
-        imageBuffer: albumArt,
+          name: 'front cover'
+        }
       },
+      title: result.trackName
     };
+  }
+
+  private async fetchAlbumArt(url: string): Promise<Buffer> {
+    return axios
+      .get(url, { responseType: 'arraybuffer' })
+      .then((response) => Buffer.from(response.data as string, 'binary'))
+      .catch(() => {
+        throw new YtdlMp3Error('Failed to fetch album art from endpoint: ' + url);
+      });
   }
 
   private async fetchResults(): Promise<SearchResult[]> {
@@ -94,14 +103,5 @@ export class SongTagsSearch {
       }
     }
     throw new YtdlMp3Error('End of results');
-  }
-
-  private async fetchAlbumArt(url: string): Promise<Buffer> {
-    return axios
-      .get(url, { responseType: 'arraybuffer' })
-      .then((response) => Buffer.from(response.data as string, 'binary'))
-      .catch(() => {
-        throw new YtdlMp3Error('Failed to fetch album art from endpoint: ' + url);
-      });
   }
 }