+++ src/__legacy/index.ts
@@ -1,4 +1,4 @@
-export { Downloader, DownloaderOptions } from './Downloader';
+export { Downloader, type DownloaderOptions } from './Downloader';
 export { FormatConverter } from './FormatConverter';
 export * from './SongTagsSearch';
 export * from './main';