+++ tsup.config.js
@@ -1,9 +1,0 @@
-import { defineConfig } from 'tsup';
-
-export default defineConfig({
-  entry: ['src/index.ts'],
-  clean: true,
-  dts: true,
-  format: ['cjs', 'esm'],
-  sourcemap: true
-});